Fixed VS warnings

+++ EntityManager.h
@@ -27,7 +27,7 @@ public:
 	bool InsertEntity(const int index, std::shared_ptr<GameEntity> entity);
 	void AddEntity(std::shared_ptr<GameEntity> entity);
 
-	int NumEntities() {	return l_entities.size(); }
+	int NumEntities() {	return static_cast<int>(l_entities.size()); }
 
 	void const DrawEntities(bool prepareMat = true);
 	void UpdateEntities(float dt);
